pyproject.toml
[build-system]
  requires = ["hatchling"]
  build-backend = "hatchling.build"

[project]
  name = "versecbot"
  version = "0.2"
  authors = [{ name = "Museus", email = "versecbot@museus.dev" }]
  description = "Extendable Discord Bot to manage common moderation tasks"
  readme = "README.md"
  requires-python = ">=3.8"
  classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
  ]
  dependencies = ["discord-py>=2,<3", "pydantic_settings>=2,<3"]

  license = "MIT"
  license-files = ["LICEN[CS]E*"]

  [project.urls]
    Homepage = "https://github.com/Museus/VerSecBot"
    Issues = "https://github.com/Museus/VerSecBot/issues"
